refactor: 상품 목록 패키지 분리
