+++ app/src/main/java/woowacourse/shopping/ui/productlist/PreviewableAsyncImage.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.ui.productlist
 
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -7,6 +7,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.res.painterResource
 import coil3.compose.AsyncImage
+import woowacourse.shopping.R
 
 @Composable
 fun PreviewableAsyncImage(
