+++ app/src/main/java/woowacourse/shopping/ui/productlist/SingleProductItem.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.ui.productlist
 
 import android.icu.text.DecimalFormat
 import androidx.compose.foundation.layout.Column
